+++ tower-async/src/lib.rs
@@ -8,6 +8,10 @@
 #![allow(elided_lifetimes_in_paths, clippy::type_complexity)]
 #![cfg_attr(test, allow(clippy::float_cmp))]
 #![cfg_attr(docsrs, feature(doc_auto_cfg, doc_cfg))]
+
+#![cfg_attr(feature = "nightly", allow(incomplete_features))]
+#![cfg_attr(feature = "nightly", feature(associated_type_bounds, return_type_notation))]
+
 // `rustdoc::broken_intra_doc_links` is checked on CI
 
 //! `async fn(Request) -> Result<Response, Error>`
@@ -244,6 +248,10 @@ pub mod layer;
 #[cfg(feature = "util")]
 #[doc(inline)]
 pub use self::util::{service_fn, ServiceExt};
+
+#[cfg(all(feature = "util", feature = "nightly"))]
+#[doc(inline)]
+pub use self::util::NightlyServiceExt;
 
 #[doc(inline)]
 pub use crate::builder::ServiceBuilder;