+++ tower-async-http/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.1 (December 19, 2023)
+
+- Update Axum example to make use of version `axum == 0.7.2`;
+
 ## 0.2.0 (November 20, 2023)
 
 - Update to http-body 1.0;